# RT_ComOps-openapi.yaml
openapi: "3.1.0"
info:
  title: "RT_ComOps API"
  description: "RT_ComOps API"
  version: "1.0.0"
servers:
  - url: "https://RT_ComOps"
paths:
  /actors/me:
    put:
      summary: "PUT actors/me"
      operationId: "updateProfile"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessActor"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BusinessActor"
    get:
      summary: "GET actors/me"
      operationId: "getMyProfile"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BusinessActor"
...
  /organizations/{id}:
    patch:
      summary: "PATCH organizations/{id}"
      operationId: "updateOrganization"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Organization"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Organization"
  /organizations:
    post:
      summary: "POST organizations"
      operationId: "createOrganization"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Organization"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Organization"
...
components:
  schemas:
    Instant:
      type: "object"
      properties: { }
    BusinessActor:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
          nullable: true
        name:
          type: "string"
          nullable: true
        businessId:
          type: "string"
          nullable: true
        niu:
          type: "string"
          nullable: true
        tradeRegistryNumber:
          type: "string"
          nullable: true
        website:
          type: "string"
          nullable: true
        contactPhone:
          type: "string"
          nullable: true
        privateAddress:
          type: "string"
          nullable: true
        businessAddress:
          type: "string"
          nullable: true
        businessProfile:
          type: "string"
          nullable: true
        createdAt:
          nullable: true
          $ref: "#/components/schemas/Instant"
...
    Organization:
      type: "object"
      properties:
        id:
          type: "string"
          format: "uuid"
          nullable: true
        businessActorId:
          type: "string"
          format: "uuid"
          nullable: true
        code:
          type: "string"
          nullable: true
        name:
          type: "string"
          nullable: true
        serviceType:
          type: "string"
          nullable: true
        isIndividualBusiness:
          type: "boolean"
          nullable: true
        status:
          type: "string"
          nullable: true
        managerId:
          type: "string"
          format: "uuid"
          nullable: true
        email:
          type: "string"
          nullable: true
        description:
          type: "string"
          nullable: true
        logoUri:
          type: "string"
          nullable: true
        logoId:
          type: "string"
          format: "uuid"
          nullable: true
        websiteUrl:
          type: "string"
          nullable: true
        socialNetwork:
          type: "string"
          nullable: true
        businessRegistrationNumber:
          type: "string"
          nullable: true
        taxNumber:
          type: "string"
          nullable: true
        capitalShare:
          type: "number"
          format: "double"
          nullable: true
        ceoName:
          type: "string"
          nullable: true
        yearFounded:
          type: "integer"
          format: "int32"
          nullable: true
        foundingDate:
          type: "string"
          format: "date"
          nullable: true
        legalForm:
          type: "string"
          nullable: true
        keywords:
          type: "string"
          nullable: true
        numberOfEmployees:
          type: "integer"
          format: "int32"
          nullable: true
        isActive:
          type: "boolean"
          nullable: true
        createdAt:
          nullable: true
          $ref: "#/components/schemas/Instant"
        updatedAt:
          nullable: true
          $ref: "#/components/schemas/Instant"
...